abcd

+++ backend/controllers/formController.js
@@ -7,17 +7,8 @@ const QuarterDetails = require("../models/quarterDetails");
 const FormDetails = require("../models/formDetails");
 const APIFeatures = require("../utils/apiFeatures");
 const moment = require('moment-timezone');
-const currentT =()=> {
-  const istTimeZone = 'Asia/Kolkata';
-  
-const currentIST = moment().tz(istTimeZone);
+const Admin = require('../models/admin');
 
-
-const currentISTString = currentIST.format('YYYY-MM-DDTHH:mm');
-
-return currentISTString;
-
-}
 
 exports.acceptForm = catchAsyncError(async (req, res, next) => {
   const formDetails = await FormDetails.findById("admin");
@@ -58,29 +49,125 @@ exports.acceptForm = catchAsyncError(async (req, res, next) => {
       maritalStatus,
       dateOfMarriage,
       contact,
+      guide,
+      tenureCompleted,
+       remarks
     } = req.body;
-
+    console.log(JSON.stringify(req.body));
     const user = await User.findById(req.user.id);
     const email = user.email;
 
     var RollNo = email.substring(0, email.indexOf("@"));
-    const form = await ResearchScholarForm.create({
-      RollNo,
-      Name,
-      Department,
-      dateOfJoining,
-      category,
-      fellowship,
-      nameOfInstitute,
-      typeOfInstitute,
-      scaleOfPayAndBasicPay,
-      presentResidentialAddress,
-      permanentAddress,
-      maritalStatus,
-      dateOfMarriage,
-      contact,
+    const form = await ResearchScholarForm.findOneAndUpdate(
+      { RollNo },
+      {
+        Name,
+        Department,
+        dateOfJoining,
+        category,
+        fellowship,
+        nameOfInstitute,
+        typeOfInstitute,
+        scaleOfPayAndBasicPay,
+        presentResidentialAddress,
+        permanentAddress,
+        maritalStatus,
+        dateOfMarriage,
+        contact,
+        guide,
+        tenureCompleted,
+        remarks
+      },
+      { upsert: true, new: true } 
+    );
+
+    res.status(200).json({
+      success: true,
+      form,
+    });
+  }
+});
+exports.form1PreSubmission = catchAsyncError(async (req, res, next) => {
+  const formDetails = await FormDetails.findById("admin");
+  const istTimeZone = 'Asia/Kolkata';
+
+  const currentTime = moment().tz(istTimeZone); // Get the current time in IST timezone
+
+  const startTime = moment.tz(formDetails.startTime, 'YYYY-MM-DDTHH:mm', istTimeZone);
+  const endTime = moment.tz(formDetails.endTime, 'YYYY-MM-DDTHH:mm', istTimeZone);
+
+  console.log("Form start time:", startTime.format('YYYY-MM-DDTHH:mm'));
+  console.log("Form end time:", endTime.format('YYYY-MM-DDTHH:mm'));
+  console.log("Current time:", currentTime.format('YYYY-MM-DDTHH:mm'));
+
+  if (currentTime.isBefore(startTime)) {
+    res.status(200).json({
+      success: false,
+      error: "Form filling has not started yet.",
+    });
+  } else if (currentTime.isAfter(endTime)) {
+    res.status(200).json({
+      success: false,
+      error: "Form filling has ended.",
+    });
+  } else {
+   
+   
+
+    const user = await User.findById(req.user.id);
+    const email = user.email;
+
+    var rn = email.substring(0, email.indexOf("@"));
+    const form = await ResearchScholarForm.findOne({RollNo:rn});
+    res.status(200).json({
+      success: true,
+      form,
     });
 
+  }
+});
+
+exports.getFormsData = catchAsyncError(async (req, res, next) => {
+  const form = await ResearchScholarForm.find();
+  res.status(200).json({
+    success: true,
+    form,
+  });
+});
+
+exports.fetchFormData = catchAsyncError(async (req, res, next) => {
+  const formDetails = await FormDetails.findById("admin");
+  const istTimeZone = 'Asia/Kolkata';
+
+  const currentTime = moment().tz(istTimeZone); // Get the current time in IST timezone
+
+  const startTime = moment.tz(formDetails.startTime, 'YYYY-MM-DDTHH:mm', istTimeZone);
+  const endTime = moment.tz(formDetails.endTime, 'YYYY-MM-DDTHH:mm', istTimeZone);
+
+
+
+  if (currentTime.isBefore(startTime)) {
+    res.status(200).json({
+      success: false,
+      error: "Form filling has not started yet.",
+    });
+  } else if (currentTime.isAfter(endTime)) {
+    res.status(200).json({
+      success: false,
+      error: "Form filling has ended.",
+    });
+  } else {
+   
+    const user = await User.findById(req.user.id);
+
+    const email = user.email;
+    
+  
+    var RollNo = email.substring(0, email.indexOf("@"));
+
+    const form =ResearchScholarForm.find({RollNo});
+    
+  
     res.status(200).json({
       success: true,
       form,
@@ -136,13 +223,14 @@ exports.acceptQuarterForm = catchAsyncError(async (req, res, next) => {
       applicationType,
       scOrST,
       occupationDate,
+      quarterPresentlyAllocated,
+      remarks,
       priorityChoices,
     } = req.body;
 
     const user = await User.findById(req.user.id);
     const email = user.email;
-    const form = await Quarter.create({
-      email,
+    const form = await Quarter.findOneAndUpdate({ email},{
       name,
       staffNumber,
       designation,
@@ -157,7 +245,45 @@ exports.acceptQuarterForm = catchAsyncError(async (req, res, next) => {
       applicationType,
       scOrST,
       occupationDate,
+      quarterPresentlyAllocated,
+      remarks,
       priorityChoices,
+    }, { upsert: true, new: true } );
+
+    res.status(200).json({
+      success: true,
+      form,
+    });
+  }
+});
+
+exports.fetchQuarterForm = catchAsyncError(async (req, res, next) => {
+  const formDetails = await QuarterDetails.findById("admin");
+  const istTimeZone = 'Asia/Kolkata';
+
+  const startTime = moment.tz(formDetails.startTime, istTimeZone);
+  const endTime = moment.tz(formDetails.endTime, istTimeZone);
+  const currentTime = moment().tz(istTimeZone);
+
+ 
+
+  if (currentTime.isBefore(startTime)) {
+    res.status(200).json({
+      success: false,
+      error: "Quarter form filling has not started yet.",
+    });
+  } else if (currentTime.isAfter(endTime)) {
+    res.status(200).json({
+      success: false,
+      error: "Quarter form filling has ended.",
+    });
+  } else {
+  
+    
+    const user = await User.findById(req.user.id);
+    const email = user.email;
+    const form = await Quarter.findOne({
+      email
     });
 
     res.status(200).json({
@@ -196,6 +322,7 @@ exports.getQuarterFormsData = catchAsyncError(async (req, res, next) => {
       ApplicationType: form.applicationType,
       ScOrST: form.scOrST,
       OccupationDate: form.occupationDate,
+      quarterPresentlyAllocated: form.quarterPresentlyAllocated,
       ...priorityProperties,
       SubmissionTime: form.submissionTime,
     };
@@ -226,13 +353,40 @@ exports.getSubmissionCount1 = catchAsyncError(async (req, res, next) => {
 });
 
 exports.deleteAllSubmission1 = catchAsyncError(async (req, res, next) => {
-  const count = await ResearchScholarForm.deleteMany();
+  const {password}= req.body;
+
+  
+  const email = req.admin.email;
+   const admin = await Admin.findOne({email}).select('+password');
+    if(!admin){
+        return next(new ErrorHandler('Wrong Password!',401));
+    }
+
+    const isPasswordMatched = await admin.comparePassword(password);
+    if(!isPasswordMatched)
+    {
+    return next(new ErrorHandler('Wrong Password!',401));
+    }
+    
+    const count = await ResearchScholarForm.deleteMany();
 
   res.status(200).json({
     success: true,
   });
 });
 exports.deleteAllSubmission2 = catchAsyncError(async (req, res, next) => {
+  const {password}= req.body;
+  const email = req.admin.email;
+  const admin = await Admin.findOne({email}).select('+password');
+    if(!admin){
+        return next(new ErrorHandler('Wrong Password!',401));
+    }
+
+    const isPasswordMatched = await admin.comparePassword(password);
+    if(!isPasswordMatched)
+    {
+    return next(new ErrorHandler('Wrong Password!',401));
+    }
   const count = await Quarter.deleteMany();
 
   res.status(200).json({
@@ -287,7 +441,7 @@ exports.deleteInForm1 = catchAsyncError(async (req, res, next) => {
   });
   
   exports.deleteInForm2 = catchAsyncError(async (req, res, next) => {
-    console.log(JSON.stringify(req.body)+"HELLO");
+    console.log(JSON.stringify(req.body));
     const data= await Quarter.findByIdAndDelete({_id:req.body._id});
     res.status(200).json({
       success: true,
